refactor(entity): editing id type. form Long to Integer.

+++ src/main/java/com/bit/healthpartnerboot/entity/RefreshToken.java
@@ -10,9 +10,9 @@ public class RefreshToken {
 
     @Id
     private String refreshToken;
-    private Long userId;
+    private Integer userId;
 
-    public RefreshToken(String refreshToken, Long userId) {
+    public RefreshToken(String refreshToken, Integer userId) {
         this.refreshToken = refreshToken;
         this.userId = userId;
     }
